[IMP] fcd_sale_order: add new button showing stock quants
