+++ fcd_sale_order/i18n/fcd_sale_order.pot
@@ -1,0 +1,47 @@
+# Translation of Odoo Server.
+# This file contains the translation of the following modules:
+# 	* fcd_sale_order
+#
+msgid ""
+msgstr ""
+"Project-Id-Version: Odoo Server 14.0\n"
+"Report-Msgid-Bugs-To: \n"
+"POT-Creation-Date: 2022-11-02 07:48+0000\n"
+"PO-Revision-Date: 2022-11-02 07:48+0000\n"
+"Last-Translator: \n"
+"Language-Team: \n"
+"MIME-Version: 1.0\n"
+"Content-Type: text/plain; charset=UTF-8\n"
+"Content-Transfer-Encoding: \n"
+"Plural-Forms: \n"
+
+#. module: fcd_sale_order
+#: model:ir.model.fields,field_description:fcd_sale_order.field_sale_order_line__display_name
+msgid "Display Name"
+msgstr ""
+
+#. module: fcd_sale_order
+#: model:ir.model.fields,field_description:fcd_sale_order.field_sale_order_line__id
+msgid "ID"
+msgstr ""
+
+#. module: fcd_sale_order
+#: model:ir.model.fields,field_description:fcd_sale_order.field_sale_order_line____last_update
+msgid "Last Modified on"
+msgstr ""
+
+#. module: fcd_sale_order
+#: model_terms:ir.ui.view,arch_db:fcd_sale_order.view_sale_order_form_inherit
+msgid "Open Stock Quant"
+msgstr ""
+
+#. module: fcd_sale_order
+#: model:ir.model,name:fcd_sale_order.model_sale_order_line
+msgid "Sales Order Line"
+msgstr ""
+
+#. module: fcd_sale_order
+#: code:addons/fcd_sale_order/models/sale_order_line.py:0
+#, python-format
+msgid "Stock Quant"
+msgstr ""
