+++ fcd_sale_order/models/__init__.py
@@ -1,1 +1,1 @@
-from . import sale_order
+from . import sale_order_line
