+++ fcd_sale_order/views/sale_order_line_views.xml
@@ -1,0 +1,21 @@
+<?xml version="1.0" encoding="utf-8"?>
+<odoo>
+
+    <record id="view_sale_order_form_inherit" model="ir.ui.view">
+        <field name="name">sale.order.form (fcd_sale_order)</field>
+        <field name="model">sale.order</field>
+        <field name="inherit_id" ref="sale.view_order_form" />
+        <field name="arch" type="xml">
+            <xpath expr="//tree/field[@name='product_uom']" position="before">
+                <button
+                    name="action_open_stock_quant"
+                    icon="fa-th-large"
+                    class="p-0"
+                    type="object"
+                    title="Open Stock Quant"
+                />
+            </xpath>
+        </field>
+    </record>
+
+</odoo>
